requirements.txt
aiobotocore==2.4.0
aiohttp==3.8.3
aiohttp-retry==2.8.3
aioitertools==0.11.0
aiosignal==1.2.0
alembic==1.8.1
amqp==5.1.1
antlr4-python3-runtime==4.9.3
aporia==2.15.0
appdirs==1.4.4
appnope==0.1.3
arrow==1.2.3
asttokens==2.1.0
async-timeout==4.0.2
asyncssh==2.12.0
atpublic==3.1.1
attrs==22.1.0
autopep8==2.0.0
backcall==0.2.0
billiard==3.6.4.0
black==22.10.0
boto3==1.24.59
botocore==1.27.59
bravado==11.0.3
bravado-core==5.17.1
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.2.0
colorama==0.4.6
commonmark==0.9.1
configobj==5.0.6
contourpy==1.0.6
cryptography==38.0.3
cycler==0.11.0
databricks-cli==0.17.3
debugpy==1.6.3
decorator==5.1.1
dictdiffer==0.9.0
diskcache==5.4.0
distro==1.8.0
docker==6.0.1
dpath==2.0.6
dulwich==0.20.50
dvc==2.34.0
dvc-data==0.25.3
dvc-http==2.27.2
dvc-objects==0.12.2
dvc-render==0.0.12
dvc-s3==2.21.0
dvc-task==0.1.4
dvclive==1.0
entrypoints==0.4
executing==1.2.0
filelock==3.8.0
Flask==2.2.2
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.1
fsspec==2022.10.0
funcy==1.17
future==0.18.2
gitdb==4.0.9
GitPython==3.1.29
grandalf==0.6
greenlet==2.0.1
gunicorn==20.1.0
hydra-core==1.2.0
idna==3.4
importlib-metadata==5.0.0
ipykernel==6.17.0
ipython==8.6.0
isoduration==20.11.0
iterative-telemetry==0.0.5
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpointer==2.3
jsonref==1.0.1
jsonschema==4.17.0
jupyter_client==7.4.4
jupyter_core==4.11.2
kiwisolver==1.4.4
kombu==5.2.4
Mako==1.2.3
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mlflow==1.30.0
mlflow-skinny==1.30.0
monotonic==1.6
moto==4.0.9
msgpack==1.0.4
multidict==6.0.2
mypy-extensions==0.4.3
mysql-connector-python==8.0.31
nanotime==0.5.2
neptune-client==0.16.12
neptune-sklearn==0.9.7
nest-asyncio==1.5.6
networkx==2.8.8
numpy==1.23.4
oauthlib==3.2.2
omegaconf==2.2.3
orjson==3.8.1
packaging==21.3
pandas==1.5.1
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.3
prometheus-client==0.15.0
prometheus-flask-exporter==0.20.3
prompt-toolkit==3.0.32
protobuf==3.20.1
psutil==5.9.3
psycopg2==2.9.5
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pycparser==2.21
pydot==1.4.2
pygit2==1.10.1
Pygments==2.13.0
pygtrie==2.5.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
pyzmq==24.0.1
querystring-parser==1.2.4
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.22.0
rfc3339-validator==0.1.4
rfc3987==1.3.8
rich==12.6.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3fs==2022.10.0
s3transfer==0.6.0
scikit-learn==1.1.3
scikit-plot==0.3.7
scipy==1.9.3
scmrepo==0.1.3
seaborn==0.12.1
shortuuid==1.0.9
shtab==1.5.7
simplejson==3.18.0
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
SQLAlchemy==1.4.43
sqlparse==0.4.3
stack-data==0.6.0
swagger-spec-validator==3.0.3
tabulate==0.9.0
tenacity==8.1.0
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
types-toml==0.10.8.1
typing_extensions==4.4.0
uri-template==1.2.0
urllib3==1.26.12
vine==5.0.0
voluptuous==0.13.1
wcwidth==0.2.5
webcolors==1.12
websocket-client==1.4.2
Werkzeug==2.2.2
wget==3.2
wrapt==1.14.1
xmltodict==0.13.0
yarl==1.8.1
yellowbrick==1.5
zc.lockfile==2.0
zipp==3.10.0
